+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
   * Switched pydineof_dineof load code to use newer `pysat.utils.io.load_netcdf`
   * Updated NEP29 and GitHub actions versions
   * Added manual test for pysat and pysatNASA Release Candidates
+  * Added manual test for pysatModels RC pip install
+  * Updated tests to new pysat and pytest standards
 * Documentation
   * Added badges and instructions for PyPi and Zenodo